Create rs_id_graph
Initial pass of SQL code.  This executes on redshift as is.  There are multiple passes of Step 7 which will be replaced with a while loop in DBT.   Steps 2 and 3 will also be changed with dbt.  Step 2 uses a listagg to concatenate SQL statements while Step 3 executes the output.  The dbt will leverage Jinja, macros, python, etc to create the SQL script as a variable (as in step 2) and then pass that variable to Step 3 for execution.