Update rs_user_journey
Changed the original page and track table selects.  In the first version, we were joining to the ID graph table on either anonymous_id or user_id matching but this was killing the performance of the query.   Changed both selects so that they only join at a single point but joined the ID Graph table twice.    This radically improved the performance of the query.  Coalesced the rudder_id's.